tests/wpt/tests/html/canvas/offscreen/text/2d.text.draw.fill.maxWidth.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.draw.fill.maxWidth.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.draw.fill.maxWidth.negative</h1>
<p class="desc">fillText handles maxWidth correctly</p>


<script>
var t = async_test("fillText handles maxWidth correctly");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.fillStyle = '#f00';
  ctx.font = '35px Arial, sans-serif';
  ctx.fillText('fail fail fail fail fail', 5, 35, -1);
  _assertGreen(ctx, 100, 50);
  t.done();

});
</script>
